[#1] feat: 게시물 전체 조회, 상세 조회, 수정, 삭제, 다중이미지 업로드 -전체 기능 작동 -Heart 변수 정리

+++ src/main/java/com/clonecode/inssagram/repository/CommentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
-    int countByPost(Post post);
+    Long countByPost(Post post);
 
     List<Comment> findAllByPostIdOrderByCreatedAtDesc(Long id);
 }
